feat(backend): Add Submission schema
I created the ReviewNote interface in an attempt to setup a virtual getter for a Submission's status, It would find the ReviewNote with the latest timestamp, ideally accessible through the document's _id and getTimestamp() method.

However, getTimestamp() is not a property of ObjectId in the type definitions for mongoose, unlike what the documentation would suggest.

Unrelated to that problem, I will be looking into the Typegoose library to avoid the TS interface / Mongoose Model duplicate definition problem.