refactor: Use mongo document types instead of custom types
Unsurprisingly in hindsight, my methods accepting a parameter of type Ref<Tournament> etc complain when provided Tournament but are perfectly happy with the MongoDB document of that Tournament.

Thus, this refactors a method that returns (and works with internally) Tournament to instead return MongoDB documents when calling code may need documents.

This also adds the TournamentDocument type to help with readability. A more generic solution to aliasing that gnarly compound type for all our documents in this way is likely possible and desirable.