refactor(rendezvous)!: Change RdvsSC generics
This breaks existing RendezvousSlashCommands. They must use the updated generic parameters scheme.

It may be best to use the TypeScript suggested approach rather than "ignoring" the error (which I'm doing because I know my solution will error when it is supposed to error). I just wasn't sure a priori if that would achieve the intended behavior.